requirements.txt
functions-framework
google-cloud-secret-manager
beautifulsoup4
requests